Remove (most) commented out code
The code removed was mainly marked for debugging or was a `cout` or
`Rcout` that had been commented out. This code has been stored in the
debug branch. To use the debugging code again, just run the following
commands

```
git checkout debug
git merge master
```

And the debug branch will be up to date with the master branch but
include the commented out code