Remove the Rhdf5lib paths from the linking step
In the past, when Rhdf5lib was included in the LinkingTo section of
the DESCRIPTION, we also had to manually specify the search paths for
the library in Makevars.in and Makevars.win. When the dependency on
Rhdf5lib was removed, these manual additions to the linker search
path were not removed. This did not fail for a long time since
these libraries were still available even though we didn't
explicitly depend on Rhdf5lib. Somehow an update to a different
dependency broke this and so this patch removes these paths
that are searching for a library that is no longer there (and
is not needed in the first place).