Fix bug with death-rebirth
When re-factoring the logic involved in a Death update step, so that
some of it was handled by the DataModel instead of the Sampler, a bug was
introduced. This bug resulted in the rebirth attempt of the Death update
step being rejected if the gibbs sampling step failed, rather than
attempting the rebirth with the original atom's mass. This commit pulls
out the logic and implements the correct Death step entirely in the
Sampler.