replace getPatternGeneSet example in vignette
Replacement of getPatternHallmarks in vignettes now that the function had changed names. Currently uses a placeholder example of testing two hallmark gene sets known to be enriched in pattern 7. The full gene sets are written in the code executed and should be replaced with a direct query of the gene sets from online or a hosted gene set in the package. I would avoid reuse of msigdbr because dependancy on this unstable package was the impetus for these changes in the first place.