replace pathway column with gene.set
Bug found in issue #107. Creating a separate column for "gene.set" apart from the "pathways" column generated by the fgsea and fora functions led to gene set names being associated with the wrong rows. To avoid this kind of issue of redundant columns naming the gene sets that caused this bug, the pathways column is renamed to gene.set then converted to a factor ordered by adjusted p-value.